feat(native)!: rename file TurboModule methods and extract files helpers
