+++ android/src/main/java/com/sherpaonnx/SherpaOnnxContentUriUtils.kt
@@ -1,0 +1,37 @@
+package com.sherpaonnx
+
+import android.content.ContentResolver
+import android.net.Uri
+import android.provider.DocumentsContract
+import java.io.InputStream
+import java.io.OutputStream
+
+/** Shared SAF/document helpers for TTS save-to-SAF and [SherpaOnnxFilesHelper]. */
+internal object SherpaOnnxContentUriUtils {
+  fun createDocumentInDirectory(
+    resolver: ContentResolver,
+    directoryUri: Uri,
+    filename: String,
+    mimeType: String
+  ): Uri {
+    return if (DocumentsContract.isTreeUri(directoryUri)) {
+      val documentId = DocumentsContract.getTreeDocumentId(directoryUri)
+      val dirDocUri = DocumentsContract.buildDocumentUriUsingTree(directoryUri, documentId)
+      DocumentsContract.createDocument(resolver, dirDocUri, mimeType, filename)
+        ?: throw IllegalStateException("Failed to create document in tree URI")
+    } else {
+      DocumentsContract.createDocument(resolver, directoryUri, mimeType, filename)
+        ?: throw IllegalStateException("Failed to create document in directory URI")
+    }
+  }
+
+  fun copyStream(inputStream: InputStream, outputStream: OutputStream) {
+    val buffer = ByteArray(8192)
+    var bytes = inputStream.read(buffer)
+    while (bytes >= 0) {
+      outputStream.write(buffer, 0, bytes)
+      bytes = inputStream.read(buffer)
+    }
+    outputStream.flush()
+  }
+}
